(hel-401) move compte de resultat dans et ms
