+++ src/frontend/ui/établissement-territorial-médico-social/bloc-budget-et-finances/BlocBudgetEtFinancesMédicoSocial.tsx
@@ -2,13 +2,13 @@ import { Bloc } from "../../commun/Bloc/Bloc";
 import { useDependencies } from "../../commun/contexts/useDependencies";
 import { BlocIndicateurVide } from "../../commun/IndicateurGraphique/BlocIndicateurVide";
 import { IndicateurGraphique } from "../../commun/IndicateurGraphique/IndicateurGraphique";
-import { GraphiqueCompteDeResultat } from "../../indicateur-métier/compte-de-resultat/GraphiqueCompteDeResultat";
 import { ContenuFondDeRoulementNetGlobal } from "../InfoBulle/ContenuFondDeRoulementNetGlobal";
 import { ContenuMontantDeLaContributionAuxFraisDeSiège } from "../InfoBulle/ContenuMontantDeLaContributionAuxFraisDeSiège";
 import { ContenuRésultatNetComptable } from "../InfoBulle/ContenuRésultatNetComptable";
 import { ContenuTauxDeCaf } from "../InfoBulle/ContenuTauxDeCaf";
 import { ContenuTauxDeVétustéConstruction } from "../InfoBulle/ContenuTauxDeVétustéConstruction";
 import styles from "./BlocBudgetEtFinancesMédicoSocial.module.css";
+import { GraphiqueCompteDeResultat } from "./compte-de-resultat/GraphiqueCompteDeResultat";
 import { ÉtablissementTerritorialBudgetEtFinancesMédicoSocialViewModel } from "./ÉtablissementTerritorialBudgetEtFinancesMédicoSocialViewModel";
 
 type BlocBudgetEtFinancesMédicoSocialProps = Readonly<{
